Delete schedules after execution
When access is granted, an EventBridgeSchedule is created to fire when
access should be revoked.  However the schedules are not deleted, and so
the list of schedules continues to grow, for ever.

This has the knock-on effect of making the revoke lambda run slower
and slower, as it has to paginate through more and more old schedules.

This commit sets the ActionAfterCompletion property on the schedule so
that once it has executed AWS will delete it.  This means that the
number of schedules should remain fairly constant over a given period of
time.

Note that I have not included anything here to delete old schedules, so
tidying-up an existing installation will be something that each team
will need to perform manually.